First attempt at including backtraces in parse errors
This required splitting parse errors up and
giving them a seperate `SimpleParseErrorKind`.
Then we can put both the backtrace and location
in the wrapper `SimpleParseError`.
This is a lot cleaner, since we no longer need
to store the location in each enum variant.
Constructing the wrapper automatically detects the backtrace,
making it appear that the only thing it needs is location and kind.

This mimics failure's support and deep down we rely
on them for backtraces.
If this attempt is succesfull,
we can get rid of that horrible panicking nonsense.